Move client and client_pool to quickwit-search.
