+++ quickwit-cluster/src/lib.rs
@@ -18,8 +18,6 @@
 //  You should have received a copy of the GNU Affero General Public License
 //  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-pub mod client;
-pub mod client_pool;
 pub mod cluster;
 pub mod service;
 pub mod utils;
