+++ quickwit-search/src/client_pool.rs
